requirements.txt
streamlit
google-generativeai
fpdf
SpeechRecognition
gTTS
pydub
